// generated/src/app/api-models/dddw-hdr.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class DddwHdr {
  dwName: string;
  columnName: string;
  dwRef: string;
  colRef: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}